tests and authentication, fixed auth token?

+++ api/routers/schoolsRouter.js
@@ -9,7 +9,7 @@ const ware = require('../middleware.js');
 
 router.route('/')
     .get(get)
-    .post(add)
+    .post(ware.authenticate, add)
     ;
 
 router.route('/:id')
@@ -26,7 +26,7 @@ function get(req, res) {
         .then(result => {
             res.status(200).json(result);
         })
-        .catch(err => res.status(500).json({ msg: 'name must be unique', err }));
+        .catch(err => res.status(500).json({ msg: 'cant find table', err }));
 }
 function getId(req, res) {
     const id = req.params.id;
@@ -65,30 +65,64 @@ function add(req, res) {
                     })
             }).catch(err => res.status(405).json({ msg: 'name must be unique', err }));
     } else {
-        res.status(422).json('Must include name, address, and funds');
+        res.status(422).json('Must include name, address, and requested_funds');
     }
 }
 
 function remove(req, res) {
     const id = req.params.id;
 
-    db.remove(id)
-        .then(result => {
-            res.status(200).json(result);
+    db.get()
+        .then(schools => {
+            let found = false;
+            for (let i = 0; i < schools.length; i++) {
+                if (schools[i].id == id) {
+                    found = true;
+                    break;
+                }
+            }
+            if (found) {
+                db.remove(id)
+                    .then(result => {
+                        res.status(200).json({ deleted: (!!result) });
+                    })
+                    .catch(err => res.status(400).json({ msg: 'id not found', err }))
+            } else {
+                res.status(404).json({ msg: 'school with id not found' });
+            }
         })
-        .catch(err => res.status(400).json({ msg: 'id not found', err }))
+
+
 }
 
 function update(req, res) {
     const { id } = req.params;
     const changes = req.body;
-    // if ((changes.name !== undefined)) {
-    db.update(id, changes)
-        .then(result => {
-            res.status(200).json(result);
+
+    db.get()
+        .then(schools => {
+            let found = false;
+            for (let i = 0; i < schools.length; i++) {
+                if (schools[i].id == id) {
+                    found = true;
+                    break;
+                }
+            }
+            if (found) {
+                if (changes) {
+                    db.update(id, changes)
+                        .then(result => {
+                            res.status(200).json(result);
+                        })
+                        .catch(err => res.status(500).json({ msg: 'could not update', err }))
+                } else {
+                    res.status(500).json({ msg: "you must give a changed school object" })
+                }
+            } else {
+                res.status(404).json({ msg: 'school with id not found' });
+            }
         })
-        .catch(err => res.status(500).json({ msg: 'could not update', err }))
-    // } else {
-    //     res.status(500).json({ msg: "you must give a school name" })
-    // }
+
+
+
 }
